ADK_RAG_Wiki_Assistant/requirements.txt
google-adk
python-dotenv
requests
beautifulsoup4
openai
numpy
